requirements.txt
streamlit==1.36.0
google-generativeai==0.7.1 # Keep for consistency, though not core to this specific project
python-dotenv==1.0.1
Pillow==10.4.0
rembg[cpu]==2.0.50 # Installs rembg with CPU support. [gpu] for GPU if you have CUDA.
onnxruntime==1.18.0 # Recommended for rembg performance